requirements.txt
pip==24.0
PyYAML==6.0.1
setuptools==65.5.0
scikit-learn==1.2.2
xgboost==2.0.0
zenml==0.55.5
zenml[server]==0.55.5
joblib==1.2.0
jupyter==1.0.0
typing-extensions==4.10.0
pandas==1.5.2
numpy==1.23.5
mlflow==2.11.3
